Fix semantics and aria roles for top info alert section
Ensure outer div is section, with role "note".
The inner div where the alert is, is an aside, with role "status" because it's not important enough to be an "alert" that needs immediate attention.
Added an info icon, but this really caused a lot of trouble. It is working now with a lot of various flexbox settings!
Set max width to 90%.
Made sure wrapping works well.

Fixes: #7