Migrate privacy pages
Current site has good privacy pages, but they need to be migrated to the new site, so that is now done.

Set a category and tags, for future archive page.

Remove the hard coded contact area from bottom of page, then set contact to true, so that the contact section is shown.

The id links the translations.