Enable section search
Debugged pagefind, and it needed a different config object. I had the old one. Switched to be a child of ui, and added the setting to showSubResults.

Added id's and anchor tags to each section on the landing, so that they now appear in search.

Refactored certs and partners, so that they too have an h2 (formatted small) and can also appear in search.