Enhances UI with fade-in animations
Improves visual appeal by adding subtle fade-in effects to elements on page load.

This change introduces fade-in animations for the mobile menu links and the hero image, creating a smoother and more engaging user experience.
It also includes other animation utility classes.

Note it did NOT work to use tailwind.config.js. You have to add the vars to your theme, and add keyframes and classes. Now it works, though!